Allow custom tile grid / projection
A function for a customized projection of input coordinates to the unit space
can be added via options. This function takes x,y and returns an object with
the transformed coordinates. It can e.g. consider a extent or even flip the
axes. It's important to note that the origin of the unit sphere is top left!